Move Banner generation to internal/ui
