+++ internal/ui/banner.go
@@ -1,0 +1,16 @@
+package ui
+
+import "strings"
+
+// Banner returns the ASCII-art banner with the k6 logo
+func Banner() string {
+	banner := strings.Join([]string{
+		`         /\      Grafana   /‾‾/  `,
+		`    /\  /  \     |\  __   /  /   `,
+		`   /  \/    \    | |/ /  /   ‾‾\ `,
+		`  /          \   |   (  |  (‾)  |`,
+		` / __________ \  |_|\_\  \_____/ `,
+	}, "\n")
+
+	return banner
+}
